backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
cx_Oracle==8.3.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
slowapi==0.1.9
sqlalchemy==2.0.23
aiosqlite==0.19.0